Move coins to utils file
