+++ rust/src/lib.rs
@@ -11,6 +11,7 @@ use cbor_event::Special as CBORSpecial;
 pub mod address;
 pub mod crypto;
 pub mod fees;
+pub mod utils;
 #[macro_use]
 pub mod prelude;
 pub mod serialization;
@@ -18,6 +19,7 @@ pub mod serialization;
 use address::*;
 use crypto::*;
 use prelude::*;
+use utils::*;
 
 #[wasm_bindgen]
 #[derive(Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
@@ -42,58 +44,6 @@ impl UnitInterval {
         Self {
             index_0: index_0,
             index_1: index_1,
-        }
-    }
-}
-
-// Specifies an amount of ADA in terms of lovelace
-// String functions are for environments that don't support u64 or BigInt/etc
-#[wasm_bindgen]
-#[derive(Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd)]
-pub struct Coin(u64);
-
-to_from_bytes!(Coin);
-
-#[wasm_bindgen]
-impl Coin {
-    // May not be supported in all environments as it maps to BigInt with wasm_bindgen
-    pub fn new(value: u64) -> Coin {
-        Self(value)
-    }
-    pub fn unwrap(&self) -> u64 {
-        self.0
-    }
-
-    // Create a Coin from a standard rust string representation
-    pub fn from_str(string: &str) -> Result<Coin, JsValue> {
-        string.parse::<u64>()
-            .map_err(|e| JsValue::from_str(&format! {"{:?}", e}))
-            .map(Coin)
-    }
-
-    // String representation of the Coin value for use from environments that don't support BigInt
-    pub fn to_str(&self) -> String {
-        format!("{}", self.0)
-    }
-
-    pub fn checked_mul(&self, other: &Coin) -> Result<Coin, JsValue> {
-        match self.0.checked_mul(other.0) {
-            Some(value) => Ok(Coin(value)),
-            None => Err(JsValue::from_str("overflow")),
-        }
-    }
-
-    pub fn checked_add(&self, other: &Coin) -> Result<Coin, JsValue> {
-        match self.0.checked_add(other.0) {
-            Some(value) => Ok(Coin(value)),
-            None => Err(JsValue::from_str("overflow")),
-        }
-    }
-
-    pub fn checked_sub(&self, other: &Coin) -> Result<Coin, JsValue> {
-        match self.0.checked_sub(other.0) {
-            Some(value) => Ok(Coin(value)),
-            None => Err(JsValue::from_str("underflow")),
         }
     }
 }
