+++ rust/src/serialization.rs
@@ -52,21 +52,6 @@ impl DeserializeEmbeddedGroup for UnitInterval {
     }
 }
 
-impl cbor_event::se::Serialize for Coin {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_unsigned_integer(self.0)
-    }
-}
-
-impl Deserialize for Coin {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        match raw.unsigned_integer() {
-            Ok(value) => Ok(Self(value)),
-            Err(e) => Err(DeserializeError::new("Coin", DeserializeFailure::CBOR(e))),
-        }
-    }
-}
-
 impl cbor_event::se::Serialize for Transaction {
     fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>) -> cbor_event::Result<&'se mut Serializer<W>> {
         serializer.write_array(cbor_event::Len::Len(3))?;
